chore: experiment with some more tracing
