+++ apps/client/src/App.tsx
@@ -1,21 +1,12 @@
-import type { UserIdentity } from "@mp/auth-client";
 import { AuthContext, createAuthClient } from "@mp/auth-client";
 import { QueryClientProvider } from "@tanstack/solid-query";
 import { Router } from "@solidjs/router";
-import type { FaroUser } from "@mp/metrics/client";
-import {
-  faroLoggerHandler,
-  getWebInstrumentations,
-  initializeFaro,
-  TracingInstrumentation,
-} from "@mp/metrics/client";
-import { createEffect, onCleanup, useContext } from "solid-js";
 import Layout from "./ui/Layout";
 import { routes } from "./routes";
 import { createQueryClient } from "./clients/query";
 import { giveAuthClientToTRPC } from "./clients/trpc";
 import { env } from "./env";
-import { LoggerContext } from "./logger";
+import { createFaroClient, useFaroIntegration } from "./clients/faro";
 
 // This is effectively the composition root of the application.
 // It's okay to define instances in the top level here, but do not export them.
@@ -23,26 +14,14 @@ import { LoggerContext } from "./logger";
 // We initialize these here because they have significantly large 3rd party dependencies,
 // and since App.tsx is lazy loaded, this helps with initial load time.
 
-const faro = initializeFaro({
-  url: env.faro.receiverUrl,
-  app: { name: "mp-client", version: env.buildVersion },
-  instrumentations: [
-    ...getWebInstrumentations(),
-    new TracingInstrumentation({ instrumentationOptions: env.faro }),
-  ],
-});
-
 const authClient = createAuthClient(env.auth);
 const queryClient = createQueryClient();
+const faroClient = createFaroClient();
 
 giveAuthClientToTRPC(authClient);
 
 export default function App() {
-  const logger = useContext(LoggerContext);
-
-  onCleanup(logger.subscribe(faroLoggerHandler(faro)));
-  createEffect(() => faro.api.setUser(deriveFaroUser(authClient.identity())));
-
+  useFaroIntegration(faroClient, authClient);
   return (
     <AuthContext.Provider value={authClient}>
       <QueryClientProvider client={queryClient}>
@@ -50,8 +29,4 @@ export default function App() {
       </QueryClientProvider>
     </AuthContext.Provider>
   );
-}
-
-function deriveFaroUser(user?: UserIdentity): FaroUser {
-  return { id: user?.id, username: user?.name };
 }
