+++ docker/.env.dev
@@ -6,8 +6,10 @@ MP_SERVER_EXPOSE_ERROR_DETAILS=true
 MP_SERVER_DATABASE_URL=postgres://mp:mp@localhost:5432/mp
 MP_SERVER_PUBLIC_DIR=public
 MP_SERVER_BUILD_VERSION=dev
+MP_SERVER_PROXY_HOST=host.docker.internal:9999
+MP_CLIENT_PROXY_HOST=host.docker.internal:5173
 MP_CLIENT_BUILD_VERSION=dev
 MP_CLIENT_EMBED_ENV=true
 KC_PUBLIC_BASE_URL=http://localhost:8080
 KC_VERIFY_EMAIL=false
-ALLOY_MP_SERVER_ADDRESS=host.docker.internal:9999
+CADDY_RATE_LIMIT_EVENTS=999999 # essentially unlimited
