+++ docker/.env.prod
@@ -3,7 +3,9 @@
 MP_SERVER_PUBLIC_MAX_AGE=31536000
 MP_SERVER_EXPOSE_ERROR_DETAILS=false
 MP_SERVER_DATABASE_URL=postgres://mp:mp@postgres:5432/mp
+MP_SERVER_PROXY_HOST=mp-server:9999
+MP_CLIENT_PROXY_HOST=mp-client:5173
 KC_PUBLIC_BASE_URL=http://keycloak:8080
 KC_VERIFY_EMAIL=true
 KC_SMTP_AUTH=true
-ALLOY_MP_SERVER_ADDRESS=mp-server:9999
+CADDY_RATE_LIMIT_EVENTS=100
