+++ docker/.env.shared
@@ -22,6 +22,10 @@ KC_BOOTSTRAP_ADMIN_PASSWORD=admin
 KC_SMTP_FROM=no-reply@${MP_DOMAIN}
 KC_SMTP_FROM_DISPLAY_NAME=MP
 
+# Caddy
+OTEL_EXPORTER_OTLP_TRACES_ENDPOINT=http://tempo:4317
+OTEL_EXPORTER_OTLP_INSECURE=true
+
 # Application
 MP_CLIENT_DOMAIN=${MP_DOMAIN}
 MP_SERVER_DOMAIN=api.${MP_DOMAIN}
