+++ docker/.env.test
@@ -5,7 +5,9 @@ MP_SERVER_PUBLIC_MAX_AGE=31536000
 MP_SERVER_EXPOSE_ERROR_DETAILS=true
 MP_SERVER_BUILD_VERSION=test
 MP_SERVER_DATABASE_URL=postgres://mp:mp@postgres:5432/mp
+MP_SERVER_PROXY_HOST=mp-server:9999
+MP_CLIENT_PROXY_HOST=mp-client:5173
 MP_CLIENT_BUILD_VERSION=test
 KC_PUBLIC_BASE_URL=http://keycloak:8080
 KC_VERIFY_EMAIL=false
-ALLOY_MP_SERVER_ADDRESS=mp-server:9999
+CADDY_RATE_LIMIT_EVENTS=100
