+++ docker/Caddyfile
@@ -1,0 +1,55 @@
+{
+	order rate_limit before basicauth
+    email admin@{$MP_DOMAIN}
+}
+
+(common_rate_limit) {
+    rate_limit {
+        zone dynamic_zone {
+            key {http.request.remote_ip}
+            events {$CADDY_RATE_LIMIT_EVENTS}
+            window 5s
+        }
+    }
+}
+
+grafana.{$MP_DOMAIN} {
+    reverse_proxy http://grafana:3000
+}
+
+auth.{$MP_DOMAIN} {
+    tracing {
+        span keycloak
+    }
+    request_header X-Trace-Id {http.vars.trace_id}
+    reverse_proxy keycloak:8080
+}
+
+faro.{$MP_DOMAIN} {
+    reverse_proxy alloy:18273
+}
+
+{$MP_CLIENT_DOMAIN} {
+    import common_rate_limit
+    tracing {
+        span mp_client
+    }
+    request_header X-Trace-Id {http.vars.trace_id}
+    reverse_proxy {$MP_CLIENT_PROXY_HOST}
+
+    # TODO turn these headers into <meta> tags instead
+    header {
+        X-Trace-Id {http.vars.trace_id}
+        traceparent "00-{http.vars.trace_id}-{http.request.uuid}-00"
+    }
+}
+
+
+{$MP_SERVER_DOMAIN} {
+    import common_rate_limit
+    tracing {
+        span mp_server
+    }
+    request_header X-Trace-Id {http.vars.trace_id}
+    reverse_proxy {$MP_SERVER_PROXY_HOST}
+}
