--- docker/caddy/base.Caddyfile
@@ -1,11 +1,0 @@
-grafana.{$MP_DOMAIN} {
-    reverse_proxy http://grafana:3000
-}
-
-auth.{$MP_DOMAIN} {
-    reverse_proxy keycloak:8080
-}
-
-faro.{$MP_DOMAIN} {
-    reverse_proxy alloy:18273
-}
