--- docker/caddy/prod.Caddyfile
@@ -1,26 +1,0 @@
-{
-	order rate_limit before basicauth
-    email admin@{$MP_DOMAIN}
-}
-
-import base.Caddyfile
-
-(common_rate_limit) {
-    rate_limit {
-        zone dynamic_zone {
-            key {http.request.remote_ip}
-            events 100
-            window 5s
-        }
-    }
-}
-
-{$MP_CLIENT_DOMAIN} {
-    import common_rate_limit
-    reverse_proxy http://mp-client:80
-}
-
-{$MP_SERVER_DOMAIN} {
-    import common_rate_limit
-    reverse_proxy http://mp-server:9999
-}
