--- docker/caddy/test.Caddyfile
@@ -1,1 +1,0 @@
-import prod.Caddyfile
