+++ docker/grafana/alloy.alloy
@@ -1,6 +1,6 @@
 prometheus.scrape "mp_server" {
 	targets = [{
-		__address__ = env("ALLOY_MP_SERVER_ADDRESS"),
+		__address__ = env("MP_SERVER_PROXY_HOST"),
 	}]
 	forward_to      = [prometheus.remote_write.default.receiver]
 	job_name        = "mp_server"
@@ -161,12 +161,16 @@ faro.receiver "mp_client" {
 
 	output {
 		logs   = [loki.process.mp_client.receiver]
-		traces = [otelcol.exporter.otlphttp.default.input]
+		traces = [otelcol.exporter.otlp.default.input]
 	}
 }
 
-otelcol.exporter.otlphttp "default" {
+otelcol.exporter.otlp "default" {
 	client {
-		endpoint = "http://tempo:4318"
+		endpoint = env("OTEL_EXPORTER_OTLP_TRACES_ENDPOINT")
+
+		tls {
+			insecure = true
+		}
 	}
 }
