+++ packages/metrics/client.ts
@@ -7,4 +7,5 @@ export {
   getWebInstrumentations,
 } from "@grafana/faro-web-sdk";
 export { TracingInstrumentation } from "@grafana/faro-web-tracing";
+export { W3CTraceContextPropagator } from "@opentelemetry/core";
 export * from "./faroLogger";
